fix(ISSUE-185): Now input_coercer are set (and called) on an argument.
